% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBigLoaderDeps.R
\name{addBigLoaderDeps}
\alias{addBigLoaderDeps}
\title{Add CSS and JS dependencies for the bigLoaders package}
\usage{
addBigLoaderDeps()
}
\description{
Add CSS and JS dependencies for the bigLoaders package
}
