% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{useSpinner}
\alias{useSpinner}
\title{Add a spinner to a Shiny output}
\usage{
useSpinner(output)
}
\arguments{
\item{output}{A Shiny output object}
}
\value{
A modified Shiny output object with a spinner
}
\description{
Add a spinner to a Shiny output
}
